+++ formio_components_synchronizer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 16.0.2.0
+
+- Fix module v16.0 migration.
+- Add fields to the Component record, to show the (full, input) paths with keys and labels in the related Form Builder.
+- Distinguish input and layout components. Therefor add `formio.component` boolean field `input`.
+- Other improvements.
+
 ## 16.0.1.0
 
 Initial release.